WIP: refactored the bumpDeps logic to follow semver
